+++ extension/package.json
@@ -20,7 +20,7 @@
     "onStartupFinished"
   ],
   "main": "./out/extension.js",
-  "types": "./index.d.ts",
+  "types": "./out/api.d.ts",
   "contributes": {
     "configuration": {
       "title": "Verde",
@@ -63,7 +63,7 @@
         "verde.allowExtensionScripting": {
           "type": "boolean",
           "default": false,
-          "description": "Allow third-party extensions to execute Luau code inside Roblox Studio via Verde's API. Each extension is prompted for consent on first use. Disable to block all extension-initiated scripting."
+          "description": "Allow third-party extensions to execute Luau code inside Roblox Studio via Verde's API. Also requires the workspace to be trusted; untrusted workspaces always refuse. Each extension is prompted for consent on first use. Disable to block all extension-initiated scripting."
         }
       }
     },
